backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "litellm>=1.82.1",
    "markdown-pdf>=1.13.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.2",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
